backend/requirements.txt
fastapi
uvicorn[standard]
aiohttp
python-multipart
python-dotenv
